PL partition sets
The illustration shows a modified version of Figure 1 (line 178, left column). In this case, C_1(alpha) and C_1(beta) almost entirely overlap (remember that the domain extends to infinity in both x and y directions). This example of the PL model parametrized by alpha and beta doesn't satisfy the condition required by Theorem 4.4.